feat(aih): refactor workshop lesson rendering
- Introduced WorkshopLessonList component to streamline rendering of workshop lessons.
- Removed inline mapping from WorkshopListRowRenderer for cleaner structure and improved readability.
- Updated WorkshopLessonItem to accept ability and abilityStatus props for enhanced access control handling.
